refactor(agents): internalize live auth helpers
