+++ src/agents/live-auth-keys.test.ts
@@ -7,14 +7,11 @@ import { beforeAll, describe, expect, it, vi } from "vitest";
 vi.unmock("../secrets/provider-env-vars.js");
 
 let collectProviderApiKeys: typeof import("./live-auth-keys.js").collectProviderApiKeys;
-let collectAnthropicApiKeys: typeof import("./live-auth-keys.js").collectAnthropicApiKeys;
-let isAnthropicBillingError: typeof import("./live-auth-keys.js").isAnthropicBillingError;
 
 async function loadModulesForTest(): Promise<void> {
   vi.resetModules();
   vi.doUnmock("../secrets/provider-env-vars.js");
-  ({ collectAnthropicApiKeys, collectProviderApiKeys, isAnthropicBillingError } =
-    await import("./live-auth-keys.js"));
+  ({ collectProviderApiKeys } = await import("./live-auth-keys.js"));
 }
 
 beforeAll(async () => {
@@ -42,63 +39,5 @@ describe("collectProviderApiKeys", () => {
         providerEnvVars: ["XAI_API_KEY"],
       }),
     ).toEqual(["xai-live-key"]);
-  });
-});
-
-describe("collectAnthropicApiKeys", () => {
-  it("does not rotate API keys over Anthropic OAuth", () => {
-    expect(
-      collectAnthropicApiKeys({
-        env: {
-          ANTHROPIC_API_KEY: "primary-key",
-          ANTHROPIC_API_KEY_OLD: "old-key",
-          ANTHROPIC_OAUTH_TOKEN: "oauth-token",
-        },
-      }),
-    ).toEqual([]);
-  });
-
-  it("keeps API-key rotation when Anthropic OAuth is unavailable", () => {
-    expect(
-      collectAnthropicApiKeys({
-        env: {
-          ANTHROPIC_API_KEY: "primary-key",
-          ANTHROPIC_API_KEY_OLD: "old-key",
-        },
-      }),
-    ).toEqual(["primary-key", "old-key"]);
-  });
-});
-
-describe("isAnthropicBillingError", () => {
-  it("does not false-positive on plain 'a 402' prose", () => {
-    const samples = [
-      "Use a 402 stainless bolt",
-      "Book a 402 room",
-      "There is a 402 near me",
-      "The building at 402 Main Street",
-    ];
-
-    for (const sample of samples) {
-      expect(isAnthropicBillingError(sample)).toBe(false);
-    }
-  });
-
-  it("matches real 402 billing payload contexts including JSON keys", () => {
-    const samples = [
-      "HTTP 402 Payment Required",
-      "status: 402",
-      "error code 402",
-      '{"status":402,"type":"error"}',
-      '{"code":402,"message":"payment required"}',
-      '{"error":{"code":402,"message":"billing hard limit reached"}}',
-      "got a 402 from the API",
-      "returned 402",
-      "received a 402 response",
-    ];
-
-    for (const sample of samples) {
-      expect(isAnthropicBillingError(sample)).toBe(true);
-    }
   });
 });
