+++ src/agents/live-test-provider-drift.ts
@@ -14,7 +14,7 @@ import {
   isServerErrorMessage,
   isTimeoutErrorMessage,
 } from "./embedded-agent-helpers/failover-matches.js";
-import { isAnthropicBillingError, isApiKeyRateLimitError } from "./live-auth-keys.js";
+import { isApiKeyRateLimitError } from "./live-auth-keys.js";
 import { isModelNotFoundErrorMessage } from "./live-model-errors.js";
 
 type LiveProviderDriftReason =
@@ -45,6 +45,21 @@ type LiveProviderDriftOptions = {
 /** Converts arbitrary thrown values into text for provider drift matchers. */
 function liveProviderErrorText(error: unknown): string {
   return error instanceof Error ? `${error.name}: ${error.message}` : String(error);
+}
+
+function isAnthropicBillingError(message: string): boolean {
+  const lower = normalizeLowercaseStringOrEmpty(message);
+  if (
+    lower.includes("credit balance") ||
+    lower.includes("insufficient credit") ||
+    lower.includes("payment required") ||
+    (lower.includes("billing") && lower.includes("disabled"))
+  ) {
+    return true;
+  }
+  return /["']?(?:status|code)["']?\s*[:=]\s*402\b|\bhttp\s*402\b|\berror(?:\s+code)?\s*[:=]?\s*402\b|\b(?:got|returned|received)\s+(?:a\s+)?402\b|^\s*402\spayment/i.test(
+    lower,
+  );
 }
 
 /** Returns whether an error is expected live auth/account drift. */
